made renameDir take an options object, instead of a param list.  improved dryrun options

+++ library/utils.js
@@ -16,6 +16,7 @@ module.exports = {
     let konsole = helpers.getConsoleByExtension(args.sourceLocation);
     let game = helpers.getGameByFilepath(args.sourceLocation);
     let fileInfo = helpers.getFileInfo(args.sourceLocation);
+    let dryrun = args.dryrun;
 
     if (args.sortIntoFolders) {
       newPath = path.join(
@@ -31,13 +32,15 @@ module.exports = {
       );
     }
 
-    try {
-      if (!fs.existsSync(path.dirname(newPath))) {
-        fs.mkdirSync(path.dirname(newPath));
+    if(!dryrun){
+      try {
+        if (!fs.existsSync(path.dirname(newPath))) {
+          fs.mkdirSync(path.dirname(newPath));
+        }
+        fs.renameSync(args.sourceLocation, newPath);
+      } catch (e) {
+        throw new Error(`There was a generic error when moving the ROM.  Error: ${e}`);
       }
-      fs.renameSync(args.sourceLocation, newPath);
-    } catch (e) {
-      throw new Error(`There was a generic error when moving the ROM.  Error: ${e}`);
     }
 
     return newPath;
